Tests/LibWeb/Text/expected/CSSOMView/mouse-event-offset-with-nested-css-transforms.txt

offsetX: 55
offsetY: 5
